requirements.txt
python-telegram-bot
# python-dotenv # Supabase client relies on env vars directly, can be loaded by application code if needed elsewhere.
fastapi
uvicorn[standard]
jinja2
python-multipart
passlib[bcrypt] # For dashboard passcode hashing
itsdangerous # For session cookies
supabase
